pyproject.toml
[tool.poetry]
name = "sample"
version = "0.1.0"
description = "Flask and React"
authors = [""]
[tool.poetry.dependencies]
python = "^3.7, !=3.8"
pandas = ">=0.24.2"
loguru = ">=0.2.5"
sqlalchemy = "^1.3.11"
alembic = "^1.3.1"
sqlalchemy_mixins = "^1.1"
flaskwebgui = "^0.1.6"
flask-caching = "^1.8.0"
flask-sqlalchemy = "^2.4.1"
flask-migrate = "^2.5.3"
datedelta = "^1.3"
Pillow = "^7.1.1"
PyInstaller = {git = 'https://github.com/pyinstaller/pyinstaller.git', branch='develop'}
pywin32-ctypes = "^0.2.0" # WINDOWS
pefile = "^2019.4.18"
"sphinxcontrib.httpdomain" = "^1.7.0"




[tool.poetry.dev-dependencies]
black = {version = ">=18.3-alpha.0",allow-prereleases = true}
sphinx = "^2.0"
doc8 = "^0.8.0"
sphinx-autodoc-typehints = "^1.6"
"sphinxcontrib.apidoc" = "^0.3.0"
flake8 = "^3.7"
pyvan = "^0.0.5"
tox = "^3.12"
pytest = "^4.6"
sphinx-rtd-theme = "^0.4.3"
docutils = ">0.15.1"

isort = "^4.3.21"
pre-commit = "^2.3.0"
darglint = "^1.3.0"

[tool.isort]
# https://github.com/timothycrosley/isort/
known_typing = "typing,types,typing_extensions,mypy,mypy_extensions"
sections = "FUTURE,TYPING,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
include_trailing_comma = true
default_section = "FIRSTPARTY"
multi_line_output = 3
force_grid_wrap = 0
use_parentheses = true
line_length = 88
skip = "migrations/**,.venv/**,idexx_app/frontend/**,get-poetry.py"

[tool.black]
# https://github.com/psf/black
line-length = 88
target-version = ["py37"]
exclude = "(migrations|.eggs|.git|.hg|.mypy_cache|.nox|.tox|.venv|.svn|_build|buck-out|build|dist)"



[build-system]
requires = ["poetry>=0.12"]#,"briefcase"]
build-backend = "poetry.masonry.api"